ajustando cloudinary

+++ src/main/java/br/com/loveanddonateapi/service/CloudinaryService.java
@@ -1,11 +1,15 @@
 package br.com.loveanddonateapi.service;
 
+import br.com.loveanddonateapi.exception.InvalidFileException;
 import com.cloudinary.Cloudinary;
 import com.cloudinary.utils.ObjectUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MaxUploadSizeExceededException;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -17,16 +21,29 @@ public class CloudinaryService {
     @Autowired
     Cloudinary cloudinary;
 
-    public Map upload(MultipartFile multipartFile) throws IOException {
-        File file = convert(multipartFile);
-        Map result = cloudinary.uploader().upload(file, ObjectUtils.emptyMap());
-        file.delete();
-        return result;
+    public Map upload(MultipartFile multipartFile) {
+        try {
+            BufferedImage bufferedImage = ImageIO.read(multipartFile.getInputStream());
+            if (bufferedImage == null) {
+                throw new InvalidFileException("O arquivo não possui um tipo válido: PNG, JPEG, JPG.");
+            }
+            File file = convert(multipartFile);
+            Map result = cloudinary.uploader().upload(file, ObjectUtils.emptyMap());
+            file.delete();
+            return result;
+        } catch (MaxUploadSizeExceededException e) {
+            throw new MaxUploadSizeExceededException(1);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
-    public Map delete(String id) throws IOException {
-        Map result = cloudinary.uploader().destroy(id, ObjectUtils.emptyMap());
-        return result;
+    public Map delete(String id) {
+        try {
+            return cloudinary.uploader().destroy(id, ObjectUtils.emptyMap());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     private File convert(MultipartFile multipartFile) throws IOException {
